requirements.txt
absl-py==2.3.1
asgiref==3.9.1
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
Django==4.2.7
flatbuffers==25.2.10
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
keras==2.13.1
libclang==18.1.1
Markdown==3.8.2
MarkupSafe==3.0.2
numpy==1.24.3
oauthlib==3.3.1
opencv-python==4.8.1.78
opt_einsum==3.4.0
packaging==25.0
Pillow==10.0.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing_extensions==4.5.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
